requirements.txt
# Core async web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# YAML config support
PyYAML>=6.0

# BLE support (optional — only needed when running on host with Bluetooth)
bleak>=0.21.0

# GPS serial support (optional — only needed when a GPS module is attached)
pyserial>=3.5
pyserial-asyncio>=0.6

# Cloud integration (optional — install only what you need)
# boto3>=1.34.0                    # AWS IoT
# google-cloud-pubsub>=2.18.0      # GCP Pub/Sub
# azure-iot-device>=2.12.0         # Azure IoT Hub

# Raspberry Pi GPIO (optional)
# RPi.GPIO>=0.7.0                  # RPi GPIO control

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
